% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_test_functions.R
\name{check_max_consecutive}
\alias{check_max_consecutive}
\title{Check the maximum successive number for each CS condition per phase across all participants}
\usage{
check_max_consecutive(df, phase_col, which_phase, CS_col)
}
\arguments{
\item{df}{A tibble that includes all individuals' trial structures.}

\item{phase_col}{Name (without ''/"") of the column that refer to specific experimental phase (e.g., phase/phase_name)}

\item{which_phase}{A string or an integer (e.g., 1 or "avoid_conditioning").}

\item{CS_col}{Name (without ''/"") of the column that refer to CS conditions (e.g., CS_names)}
}
\description{
Check the maximum successive number for each CS condition per phase across all participants
}
\examples{
# check_max_consequtive(inner_d2, phase_name, "avoid_conditioning", CS_names)

}
