% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{check_tracker_data}
\alias{check_tracker_data}
\title{Clean up and check tracker data}
\usage{
check_tracker_data(df)
}
\arguments{
\item{df}{A trial results tibble. Must contain the columns trial_type,
walk_pos_x, walk_pos_z, head_movement_data_0}
}
\value{
A tibble with head tracker data frames cleaned up and added column
valid_head_movement_data_0
}
\description{
This function removes duplicate time stamps from tracker data frames, and
returns number of duplicates, within tracker data frames. It also checks the
validity of head tracker starting position and momentary speed using the
vrthreat package function check_valid_movement
}
