% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{find_CS}
\alias{find_CS}
\title{Determine CS+ and CS- for each subject}
\usage{
find_CS(CS, US)
}
\arguments{
\item{CS}{A string vector indicates stimulus name for each trial}

\item{US}{An integer vector indicating the US presence for each trial}
}
\value{
A list of CS- and CS+
}
\description{
Uses trial-by-trial CS and US information from a trial results data frame
to determine CS+ and CS-, ignoring trials in which no CS was presented. This
function assumes that there is just one CS+ and one CS-
}
\examples{
CS <- c('blue_cube', 'yellow_cube', 'blue_cube', 'yellow_cube')
US <- c(1,0,1,0)
find_CS(CS, US)
}
