% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{generate_outcome_variables}
\alias{generate_outcome_variables}
\title{Generate behavioral outcome variables}
\usage{
generate_outcome_variables(df)
}
\arguments{
\item{df}{A trial results tibble}
}
\value{
A trial results tibble with variables added
}
\description{
This function generates some commonly used outcome variables:
- tracker validity and max number of tracker timestamp duplicates
- CS type (CSm, CSp)
- timing information (CS onset, US onset, US offset)
- mean/min/max distance from US during CS presentation
- distance travelled during CS presentation
- maximum movement speed during CS presentation
}
