% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{normalise_data}
\alias{normalise_data}
\title{Normalise data to mean of CS- trials for each participant}
\usage{
normalise_data(data, condition)
}
\arguments{
\item{data}{A numerical vector}

\item{condition}{A factor with several levels of which one must be CSm,
indicating the CS condition}
}
\value{
A normalised data vector of the same length as data
}
\description{
Normalise data to mean of CS- trials for each participant
}
\examples{
scr <- c(3, 9, 10, 2)
cs <- c('CSm', 'CSp', 'CSp', 'CSm')
normalise_data(scr, cs)
}
