% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{parse_phases}
\alias{parse_phases}
\title{Separate experiment phases}
\usage{
parse_phases(break_before)
}
\arguments{
\item{break_before}{A logical vector (TRUE/FALSE) with trial-by-trial information}
}
\value{
a vector of phase number for each trial
}
\description{
Phases are defined by the presence of breaks before a trial. This functions
finds these breaks
}
\examples{
break_before <- c(TRUE,TRUE,FALSE,FALSE,FALSE)
parse_phases(break_before)
}
