% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_by_trial}
\alias{plot_by_trial}
\title{Draw trial-level graphs}
\usage{
plot_by_trial(
  df,
  phases,
  DV,
  cs_types = c("CSp", "CSm"),
  fill_colors = c("red", "lightgreen", "blue", "yellow")
)
}
\arguments{
\item{df}{A trial results tibble}

\item{phases}{Experimental phases}

\item{DV}{Outcome variables (e.g., 'mean_dist_CS')}

\item{cs_types}{A string vector of needed CS types (e.g., 'CS1p', 'CS1m')}

\item{fill_colors}{A string vector of fill colors for different CS conditions (e.g., 'red', 'lightgreen')}
}
\description{
At each trial, this function will plot the average value and its error bars
for both conditions (CS+/CS-) across all experimental phases
}
\examples{
library(dplyr)
library(ggplot2)

create_sub_data <- function(seed) {
set.seed(seed)
d_sub <-
 tibble::tibble(CS = factor(rep(c('CSp', 'CSm', 'CS2p', 'CS2m'), each = 10))
        \%>\% sample) \%>\%
 dplyr::mutate(
   mean_dist_CS = dplyr::case_when(
     CS == 'CSp' ~ rnorm(1,2.5,.3),
     CS == 'CSm' ~ rnorm(1,1,.3),
     CS == 'CS2p' ~ rnorm(1,3,.3),
     CS == 'CS2m' ~ rnorm(1, 2, .3)
   ),
   trial_num_all = 1:n())
return(d_sub)
}

d <- NULL
for (i in 1:20) {
 d_sub <- create_sub_data(seed = i)
 d <- rbind(d, d_sub)
}
d$pp <- as.character(rep(c(1:20), each = 40))
d <-
 d \%>\% dplyr::mutate(phase = rep(2, times = n()))
plot_by_trial(d, 2, "mean_dist_CS", cs_types = c('CSp', 'CSm', 'CS2p', 'CS2m'),
             fill_colors = c('red', 'lightgreen', 'blue', 'yellow'))

}
