% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_by_trial_con_diff}
\alias{plot_by_trial_con_diff}
\title{Plot trial-by-trial condition differences}
\usage{
plot_by_trial_con_diff(
  df,
  phases,
  DV_col,
  phase_col,
  ppid_col,
  trial_num_col,
  CS_col,
  cs_types = c("CSp", "CSm"),
  smooth = FALSE
)
}
\arguments{
\item{df}{A trial results tibble}

\item{phases}{A string vector (e.g., c(2:5)) that indicates experimental phases}

\item{DV_col}{Column name refers to the dependent variable}

\item{phase_col}{Column name refers to the specific experimental phase}

\item{ppid_col}{Column name refers to the participants' ID}

\item{trial_num_col}{Column name refers to the trial number per phase per subject}

\item{CS_col}{Column name refers to the CS stimuli}

\item{cs_types}{A string vector of needed CS types (e.g., 'CS1p', 'CS1m')}

\item{smooth}{A bolean value (TRUE or FALSE)}
}
\description{
i.e., on each trial the CS+/CS- difference, collapsed across
participants
}
\examples{
library(dplyr)
library(ggplot2)

create_sub_data <- function(seed) {
set.seed(seed)
d_sub <-
 tibble::tibble(CS = factor(rep(c('CSp', 'CSm'), each = 10))
        \%>\% sample) \%>\%
 dplyr::mutate(
   mean_dist_CS = dplyr::case_when(
     CS == 'CSp' ~ rnorm(1,2.5,.3),
     CS == 'CSm' ~ rnorm(1,.5,.3)
   ),
   trial_num_all = 1:n())
return(d_sub)
}

d <- NULL
for (i in 1:20) {
 d_sub <- create_sub_data(seed = i)
 d <- rbind(d, d_sub)
}

d$pp <- as.character(rep(c(1:20), each = 20))

d <-
 d \%>\% dplyr::mutate(phase = rep(2, times = n()))

plot_by_trial_con_diff(d, 2, mean_dist_CS, phase,
                      pp, trial_num_all, CS, cs_types = c('CSp', 'CSm'), smooth = TRUE)

}
