% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_by_trial_distance}
\alias{plot_by_trial_distance}
\title{Plot the changes in distances as time goes on at the trial level across all participants}
\usage{
plot_by_trial_distance(
  head_data,
  pp_col,
  trial_col,
  cs_col,
  cs_level = c("CSp", "CSm"),
  label_x = "",
  label_y = "",
  add_CI = FALSE,
  ref_x = 0,
  ref_z = 2.5
)
}
\arguments{
\item{head_data}{A head movement data frame/tibble}

\item{pp_col}{Column name refers to the participants' ID}

\item{trial_col}{Column name refers to the trial information}

\item{cs_col}{Column name refers to the CS stimuli}

\item{cs_level}{A string vector that refers to CS levels (e.g., 'CSp'/'CS2p', 'CSm'/'CS2m')}

\item{label_x}{A string (e.g., 'CS')}

\item{label_y}{A string (e.g., 'mean_dist_CS')}

\item{add_CI}{A bolean value (TRUE/FALSE), if TRUE then will add CI to the lines}

\item{ref_x}{An integer (e.g., 0)}

\item{ref_z}{An integer (e.g., 2.5)}
}
\description{
This function can be used for multiple/1 experimental phases
}
