% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Read and tidy all Unity output csv files from a single study}
\usage{
read_raw_data(user_dir)
}
\arguments{
\item{user_dir}{Parent directory where data are stored}
}
\value{
A trial results tibble
}
\description{
Reads the "trial_results.csv" output files from all participants in a parent
folder, updates file pointers for the referenced files, and reads all
head and hand tracker information
}
