% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_functions.R
\name{summarise_demographics_freq}
\alias{summarise_demographics_freq}
\title{Summary frequency demographic information}
\usage{
summarise_demographics_freq(df, DVs = "sex")
}
\arguments{
\item{df}{A tibble with participant information}

\item{DVs}{A string vector of demographic variables (default "sex")}
}
\value{
A summarised tibble
}
\description{
Summary frequency demographic information
}
\examples{
d <- tibble::tibble(sex = c('male', 'male', 'male',
                    rep("female", times = 4),
                    rep("non-binary", times = 5)))

summarise_demographics_freq(d)
}
