% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_functions.R
\name{summarise_outcome_variables}
\alias{summarise_outcome_variables}
\title{Summarise all outcome variables of a specific phase}
\usage{
summarise_outcome_variables(df, i_phase, DVs, cs_types = c("CSp", "CSm"))
}
\arguments{
\item{df}{A trial results tibble}

\item{i_phase}{An integer that indicates a specific experimental phase (e.g., 1)}

\item{DVs}{A string vector of outcome variables (e.g., c('mean_distance', 'maximum_distance'))}

\item{cs_types}{A string vector of expected CS types (e.g., 'CS1p', 'CS1m')}
}
\value{
A summarised tibble
}
\description{
For each outcome variable per condition (CS+/CS-), this function
computes the condition means, using summarise
}
