% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_functions.R
\name{summary_effect_size}
\alias{summary_effect_size}
\title{Compute effect sizes for a given DV of a specific phase}
\usage{
summary_effect_size(df, i_phase, DV, cs_types = c("CSp", "CSm"))
}
\arguments{
\item{df}{A trial results tibble}

\item{i_phase}{An integer or a string, which indicate a specific phase}

\item{DV}{A string (e.g., "mean distance")}

\item{cs_types}{A string vector of needed CS types (e.g., 'CS1p', 'CS1m')}
}
\value{
A list object
}
\description{
This function performs t-test and computes cohens_d and hedges_g
}
