% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_test_functions.R
\name{test_counterbalance}
\alias{test_counterbalance}
\title{Test for counterbalance}
\usage{
test_counterbalance(
  df,
  i_phase,
  cs_column,
  cs_type,
  stimulus,
  ppid,
  sub_num = 0
)
}
\arguments{
\item{df}{A tibble that include all individuals' trial structures}

\item{i_phase}{A string (e.g., 'first_avoid_conditioning') or An integer (e.g., 1)}

\item{cs_column}{Name of the column that refers to CS information}

\item{cs_type}{A string that indicates a specific CS type (e.g., 'CS1p_USm')}

\item{stimulus}{Name of the column that refers to specific stimuli}

\item{ppid}{Name of the column that refer to participants' ID}

\item{sub_num}{An integer that indicates number of participants. The default value is 0}
}
\value{
A boolean value (TRUE/FALSE)
}
\description{
This function will test for a specific phase, whether a specific
CS (e.g., CS1+/CS1-) is counterbalanced across participants
}
