% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_test_functions.R
\name{test_punish_pp}
\alias{test_punish_pp}
\title{Test the punishment probability per participant}
\usage{
test_punish_pp(df, i_phase, ppid, punish_prob, ...)
}
\arguments{
\item{df}{A tibble that include all individuals' trial structures}

\item{i_phase}{A string (e.g., 'first_avoid_conditioning') or An integer (e.g., 1)}

\item{ppid}{Name of column for participant ID}

\item{punish_prob}{Name of the column for punishment probability (binary: 0/1)}

\item{...}{Punishment probability given by researcher (e.g., 0.375)}
}
\value{
A boolean value (TRUE/FALSE)
}
\description{
Test the punishment probability per participant
}
