% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_test_functions.R
\name{test_stimuli_allocation_consistency}
\alias{test_stimuli_allocation_consistency}
\title{Check whether the stimuli allocation is consistent for the CS conditions across all experimental phases for each participant}
\usage{
test_stimuli_allocation_consistency(df, CS_col, cs_types, stimulus_col, pp_col)
}
\arguments{
\item{df}{A tibble that includes all individuals' trial structures.}

\item{CS_col}{Name (without ''/"") of the column that refer to CS conditions (e.g., CS_names)}

\item{cs_types}{A string vector for CS conditions (e.g., c('CSp_USp_A', 'CSp_USm_AC'))}

\item{stimulus_col}{Name (without ''/"") of the column that refer to stimulus names (e.g., stimulus_name or stimulus_name2)}

\item{pp_col}{Name (without ''/"") of the column that refer to participants ID}
}
\value{
A boolean value (TRUE/FALSE)
}
\description{
For example, pp.1 has two CS stimuli 'CSp_USp_A' and 'CSp_USm_AC', this function examines whether the corresponding...
stimulus_names are the same for these two CS stimuli across all experimental phases.
}
\examples{
# test_stimuli_allocation_consistentcy(d, CS_names, c("CSp_USp_BC", "CSp_USm_B", "BD"), stimulus_name, ppid)
# test_stimuli_allocation_consistentcy(d, CS_names, c("AB"), stimulus_name2, ppid)

}
