% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_test_functions.R
\name{test_stimuli_phase_pp}
\alias{test_stimuli_phase_pp}
\title{Test number of stimuli per condition per phase per participant}
\usage{
test_stimuli_phase_pp(df, i_phase, ppid, stimulus, ...)
}
\arguments{
\item{df}{A tibble that include all individuals' trial structures}

\item{i_phase}{A string (e.g., 'first_avoid_conditioning') or An integer (e.g., 1)}

\item{ppid}{Name of column for participant ID}

\item{stimulus}{Name of column for stimulus names}

\item{...}{Number of stimuli per condition that is defined by the researcher
Should be an integer (e.g., 32)}
}
\value{
A boolean value (TRUE/FALSE)
}
\description{
Test number of stimuli per condition per phase per participant
}
