% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_test_functions.R
\name{test_stimuli_platform_consistency}
\alias{test_stimuli_platform_consistency}
\title{Check the consistency of stimulus allocation across two visual platforms}
\usage{
test_stimuli_platform_consistency(
  df,
  CS_col,
  s_vector1,
  s_vector2,
  stimulus_col,
  stimulus2_col,
  pp_col
)
}
\arguments{
\item{df}{A tibble that includes all individuals' trial structures.}

\item{CS_col}{Name (without ''/"") of the column that refer to CS conditions (e.g., CS_names)}

\item{s_vector1}{A string vector for stimulus names (e.g., 'CSp_USp_BC')}

\item{s_vector2}{A string vector for stimulus names (e.g., 'CSp_USp_AB')}

\item{stimulus_col}{Name (without ''/"") of the column that refer to stimulus names (e.g., stimulus_name or stimulus_name2)}

\item{stimulus2_col}{Name (without ''/"") of the column that refer to stimulus names (e.g., stimulus_name or stimulus_name2)}

\item{pp_col}{Name (without ''/"") of the column that refer to participants ID}
}
\description{
for each participant, if one condition stimulus (e.g., A) can appear on both platforms
(i.e., stimulus_name and stimulus_name2) on different phases, this function checks
whether the allocation is consistent for both platforms
}
\examples{
# test_stimuli_platform_consistentcy(d, CS_names, c("CSp_USm_C", "CD"), c("AC", "CSp_USp_BC"), stimulus_name, stimulus_name2, ppid)


}
