% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_test_functions.R
\name{test_trial_CS_num}
\alias{test_trial_CS_num}
\title{Test the number of unique CS per trial for a specific phase}
\usage{
test_trial_CS_num(df, i_phase, trial, CS)
}
\arguments{
\item{df}{A tibble that include all individuals' trial structures}

\item{i_phase}{An integer (e.g., 1) or a string (e.g., 'first_avoid_conditioning')}

\item{trial}{The column name in the df tibble that contain trial number information}

\item{CS}{The column name in the df tibble that contain CS information}
}
\value{
A boolean value, either TRUE or FALSE
}
\description{
Test the number of unique CS per trial for a specific phase
}
