% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_test_functions.R
\name{test_trial_pp}
\alias{test_trial_pp}
\title{Test how many trials per participant have}
\usage{
test_trial_pp(df, ppid, trial)
}
\arguments{
\item{df}{A tibble that include all individuals' trial structures}

\item{ppid}{The name of the column that contains participant ID}

\item{trial}{The name of the column that contains trial information}
}
\value{
A boolean value (TRUE/FALSE)
}
\description{
Test how many trials per participant have
}
