% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{update_session_number}
\alias{update_session_number}
\title{Update session number based on experimental days per participant}
\usage{
update_session_number(exp)
}
\arguments{
\item{ppid}{A string vector of participant ids that include the day}
}
\description{
For studies that span multiple days, the ppid might include an indication of
experimental day. This function parses the ppid and extracts the day as
numerical variable
}
\examples{
exp <- c('coglearn_exp1_s3_day1', 'coglearn_exp1_s3_day2', 'coglearn_exp1_s3_day3')
update_session_number(exp)
}
